chore: check image exists in Supabase before delete

+++ src/main/java/com/sm/jeyz9/storemateapi/services/impl/ProductServiceImpl.java
@@ -227,7 +227,9 @@ public class ProductServiceImpl implements ProductService {
         try {
             product.getProductImage().stream().map(
                     p -> {
-                        supabaseService.deleteProductImage(p.getImageUrl());
+                        if(supabaseService.imageExists(p.getImageUrl())){
+                            supabaseService.deleteProductImage(p.getImageUrl());
+                        }
                         return null;
                     }
             ).toList();
